# Conflicts:
#	app/src/main/java/com/hoffhaxx/app/concurs/RecyclerViewAdapterInfoFragment.kt
#	app/src/main/java/com/hoffhaxx/app/concurs/misc/QuestRepository.kt
#	app/src/main/java/com/hoffhaxx/app/concurs/misc/SharedPreferencesRepository.kt

+++ app/src/main/java/com/hoffhaxx/app/concurs/misc/data/QuestData.kt
@@ -1,5 +1,5 @@
 package com.hoffhaxx.app.concurs.misc.data
 
-data class Quest(val description: String, val points: Double)
+data class Quest(val description: String, val points: Int)
 
 data class QuestData(val result: String, val quests: MutableList<Quest>)
